Do not polyfill VTTCue / Made it a module of VTTParser JW7-3075
