+++ test/config.js
@@ -20,9 +20,6 @@
     if (!('Promise' in window)) {
         deps.push('polyfills/promise');
     }
-    if (!('VTTCue' in window)) {
-        deps.push('polyfills/vttcue');
-    }
     if (!('console' in window) || !('log' in window.console) ) {
         window.console = {
             log: function() {}
